fix: resolve integration test hanging issue
Root cause: aiohttp sessions cannot be shared across async tasks due to
"Timeout context manager should be used inside a task" error.

Solution:
1. Created shared conftest.py with properly scoped fixtures
2. Added pytest-timeout to prevent infinite hangs (60s per test)
3. Removed duplicate client fixtures from individual test files
4. Kept function scope for fixtures to ensure each test gets fresh session

This resolves the CI issue where integration tests would stall after
the first test due to session sharing across different async tasks.
