+++ pyproject.toml
@@ -109,6 +109,7 @@ addopts = [
 markers = [
     "integration: Integration tests that require real API access",
 ]
+timeout = 60  # Per-test timeout in seconds (prevents hanging)
 
 [tool.coverage.run]
 source = ["src/pylxpweb"]
@@ -144,6 +145,7 @@ dev = [
     "pytest-asyncio>=1.3.0",
     "pytest-cov>=7.0.0",
     "pytest-rerunfailures>=15.0",
+    "pytest-timeout>=2.4.0",
     "python-dotenv>=1.2.1",
     "pyyaml>=6.0.3",
     "twine>=6.2.0",
